Move page_text_labels to a separate module
