+++ src/modules/dialogs/import_project.js
@@ -21,7 +21,7 @@
  */
 
 // eslint-disable-next-line camelcase
-import {page_text_label} from '../blockly/language/en/messages';
+import {page_text_label} from '../blockly/language/en/page_text_labels';
 import {LOCAL_PROJECT_STORE_NAME, TEMP_PROJECT_STORE_NAME} from '../constants';
 import {insertProject, isProjectChanged} from '../editor';
 import {uploadHandler, appendProjectCode} from '../editor';
